Make flow closure conform to RFC 793 TCP
A packet with a FIN but no ACK is generally considered anomalous.
Strictly speaking, the RFC 793 TCP spec does not consider a missing
ACK a reason to prevent closing a connection.  Yaf seems to take the
same approach and closes a flow for a FIN or RST flag.

We close a flow on a FIN or RST, but we track if the packet with FIN
set also sets ACK.

+++ ipaddress.go
@@ -9,7 +9,7 @@ import (
 
 // IPAddress ...
 type IPAddress struct {
-	Version int
+	Version byte
 	Address [16]byte
 }
